requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-huggingface>=0.0.3
langchain-text-splitters==0.3.0
sentence-transformers>=2.2.2
faiss-cpu
python-dotenv
httpx
tiktoken
pydantic